requirements-dev.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements-dev.in -o requirements-dev.txt
factory-boy==3.3.0
    # via -r requirements-dev.in
faker==26.0.0
    # via factory-boy
iniconfig==2.0.0
    # via pytest
packaging==24.1
    # via pytest
pluggy==1.5.0
    # via pytest
pytest==8.2.2
    # via
    #   -r requirements-dev.in
    #   pytest-django
pytest-django==4.8.0
    # via -r requirements-dev.in
python-dateutil==2.9.0.post0
    # via faker
ruff==0.5.0
    # via -r requirements-dev.in
six==1.16.0
    # via python-dateutil
